third_party/blink/web_tests/printing/pageNumerForElementById-expected.txt

Test for internals.pageNumber()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: page number of "firstPage" is 0
PASS: page number of "secondPage" is 1
PASS: page number of "thirdPage" is 2
PASS: page number of "thirdPage2" is 2
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
